chore: rename `icij_worker.task` into `icij_worker.objects`
